security: disable /api/debug/* routes in production
- implemented fail-fast security guard in proxy.ts (middleware)
- blocks all /api/debug/* routes with 404 in production
- consolidated auth and proxy header logic into proxy.ts
- added integration tests for production guard
- updated playwright tests for environment awareness
- verified build-time recognition of proxy.ts as middleware
