+++ proxy.ts
@@ -1,72 +1,68 @@
-// File: proxy.ts (NextAuth Reverse Proxy)
-/**
- * Proxy to handle reverse proxy headers for NextAuth.js.
- * This ensures that HTTPS cookies work properly behind a reverse proxy.
- */
-import { NextResponse } from 'next/server'
-import type { NextRequest } from 'next/server'
+import { withAuth, NextRequestWithAuth } from 'next-auth/middleware'
+import { NextResponse, NextFetchEvent } from 'next/server'
 
-// Base path for auth routes
 const API_AUTH_BASE = '/api/auth/'
 
-export function proxy(request: NextRequest) {
-  // Only handle auth routes
-  if (!request.nextUrl.pathname.startsWith(API_AUTH_BASE)) {
-    return NextResponse.next()
-  }
+const authMiddleware = withAuth(
+  function middleware(req: NextRequestWithAuth) {
+    if (
+      process.env.NODE_ENV === 'production' &&
+      req.nextUrl.pathname.startsWith('/api/debug')
+    ) {
+      return NextResponse.json(
+        { error: 'Endpoint unavailable in production' },
+        { status: 404 }
+      )
+    }
 
-  const response = NextResponse.next()
+    if (req.nextUrl.pathname.startsWith(API_AUTH_BASE)) {
+      const response = NextResponse.next()
+      const forwardedHost = req.headers.get('x-forwarded-host')
+      const forwardedProto = req.headers.get('x-forwarded-proto')
+      const forwardedPort = req.headers.get('x-forwarded-port')
+      const host = req.headers.get('host')
 
-  // Handle reverse proxy headers for NextAuth
-  const forwardedHost = request.headers.get('x-forwarded-host')
-  const forwardedProto = request.headers.get('x-forwarded-proto')
-  const forwardedPort = request.headers.get('x-forwarded-port')
-  const host = request.headers.get('host')
+      const actualHost = forwardedHost || host || ''
+      const actualProto = forwardedProto || 'https'
+      const actualPort = forwardedPort || ''
 
-  // Determine the actual host being accessed
-  const actualHost = forwardedHost || host || ''
-  const actualProto = forwardedProto || 'https'
-  const actualPort = forwardedPort || ''
-
-  // Reconstruct the full URL with port for NextAuth
-  if (actualHost) {
-    // Ensure Host header includes the port if not already present and port is custom
-    let hostWithPort = actualHost
-    if (actualPort && !actualHost.includes(':')) {
-      // Add port only if it's non-standard (444 for dev, or explicitly forwarded)
-      if (actualPort !== '443') {
-        hostWithPort = `${actualHost}:${actualPort}`
+      if (actualHost) {
+        let hostWithPort = actualHost
+        if (actualPort && !actualHost.includes(':')) {
+          if (actualPort !== '443') {
+            hostWithPort = `${actualHost}:${actualPort}`
+          }
+        }
+        response.headers.set('x-forwarded-host', hostWithPort)
+        response.headers.set('x-forwarded-proto', actualProto)
+        if (actualPort) {
+          response.headers.set('x-forwarded-port', actualPort)
+        }
+        if (actualProto === 'https') {
+          response.headers.set('x-forwarded-ssl', 'on')
+        }
       }
+      return response
     }
 
-    response.headers.set('x-forwarded-host', hostWithPort)
-    response.headers.set('x-forwarded-proto', actualProto)
-
-    if (actualPort) {
-      response.headers.set('x-forwarded-port', actualPort)
-    }
-
-    // Ensure NextAuth recognizes HTTPS
-    if (actualProto === 'https') {
-      response.headers.set('x-forwarded-ssl', 'on')
-    }
+    return NextResponse.next()
+  },
+  {
+    callbacks: {
+      authorized: () => true,
+    },
   }
+)
 
-  // Debug logging in development
-  if (process.env.NODE_ENV === 'development') {
-    console.log('[Proxy] Auth request:', {
-      pathname: request.nextUrl.pathname,
-      host: request.headers.get('host'),
-      forwardedHost,
-      forwardedProto,
-      forwardedPort,
-    })
-  }
-
-  return response
+export function proxy(req: NextRequestWithAuth, event: NextFetchEvent) {
+  return authMiddleware(req, event)
 }
 
 export const config = {
-  // Note: matcher must be static strings for Next.js static analysis
-  matcher: ['/api/auth/:path*'],
+  matcher: [
+    '/api/auth/:path*',
+    '/api/internal/:path*',
+    '/api/debug/:path*',
+    '/api/users/:path*',
+  ],
 }
